Move NumberField to compatibility package
