+++ touchkit-compatibility/src/test/java/org/vaadin/touchkit/v7/itest/NumberFieldTest.java
@@ -1,11 +1,11 @@
-package org.vaadin.touchkit.itest;
+package org.vaadin.touchkit.v7.itest;
 
-import org.vaadin.touchkit.AbstractTouchKitIntegrationTest;
-import org.vaadin.touchkit.ui.NumberField;
+import org.vaadin.touchkit.v7.AbstractTouchKitIntegrationTest;
+import org.vaadin.touchkit.v7.ui.NumberField;
 
+import com.vaadin.ui.Notification;
 import com.vaadin.v7.event.FieldEvents.TextChangeEvent;
 import com.vaadin.v7.event.FieldEvents.TextChangeListener;
-import com.vaadin.ui.Notification;
 import com.vaadin.v7.ui.TextField;
 
 public class NumberFieldTest extends AbstractTouchKitIntegrationTest {
@@ -16,7 +16,8 @@ public class NumberFieldTest extends AbstractTouchKitIntegrationTest {
         final NumberField nf = new NumberField("NumberField");
         nf.setImmediate(true);
         nf.addTextChangeListener(new TextChangeListener() {
-            public void textChange(TextChangeEvent event) {
+            @Override
+			public void textChange(TextChangeEvent event) {
                 Notification.show("New numberfield value: " + event.getText());
             }
         });
@@ -24,7 +25,8 @@ public class NumberFieldTest extends AbstractTouchKitIntegrationTest {
         final TextField tf = new TextField("TextField");
         tf.setImmediate(true);
         tf.addTextChangeListener(new TextChangeListener() {
-            public void textChange(TextChangeEvent event) {
+            @Override
+			public void textChange(TextChangeEvent event) {
                 nf.setValue(event.getText());
             }
         });
